chore: move endpoint to correct file
